assessment_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
sqlalchemy
pytransitions
Jinja2
periodiq